+++ test/lib/expectations.js
@@ -1,6 +1,6 @@
 const chai = require('chai');
 const cheerio = require('cheerio');
-const siteRoot = require('../../app/lib/constants').siteRoot;
+const constants = require('../../app/lib/constants');
 
 const expect = chai.expect;
 
@@ -13,7 +13,7 @@ function breadcrumbContent($) {
   expect($('.nhsuk-c-breadcrumb__item').eq(0).text().trim()).to.equal('Home');
   expect($('.nhsuk-c-breadcrumb__item').eq(0).find('a').prop('href')).to.equal('https://www.nhs.uk');
   expect($('.nhsuk-c-breadcrumb__item').eq(1).text().trim()).to.equal('Find a psychological therapies service');
-  expect($('.nhsuk-c-breadcrumb__item').eq(1).find('a').prop('href')).to.equal(siteRoot);
+  expect($('.nhsuk-c-breadcrumb__item').eq(1).find('a').prop('href')).to.equal(constants.siteRoot);
 }
 
 // eslint-disable-next-line no-script-url