+++ test/unit/lib/canonicalUrl.js
@@ -1,0 +1,17 @@
+const chai = require('chai');
+const canonicalUrl = require('../../../app/lib/canonicalUrl');
+const siteRoot = require('../../../app/lib/constants').siteRoot;
+
+const expect = chai.expect;
+
+describe('canonicalUrl', () => {
+  it('should return the current URL via https', () => {
+    const hostname = 'some.host.name';
+    const reqMock = { hostname };
+    const expectedUrl = `https://${hostname}${siteRoot}/`;
+
+    const url = canonicalUrl(reqMock);
+
+    expect(url).to.equal(expectedUrl);
+  });
+});