+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ===================
 - Update npm dependencies
 - Upgrade Docker container to use latest node LTS
+- Add additional (and update existing) metadata properties to meet minimum set
 
 0.16.0 / 2019-03-20
 ===================